Basically total rewrite of the device configuration sub-system. Now settings are stored via text keys that are human readable. And, device configuration items now can automatically be checked and saved without having to explicitly code them.
This all allows for the ability for any device to add configuration
items even as it builds upon existing base classes all the while not
having to worry about adding things to the serial console nor where
things are actually stored in EEPROM or how. MUCH easier to work with.